+++ packages/ulangi-mobile/src/delegates/spaced-repetition/SpacedRepetitionSettingsScreenDelegate.ts
@@ -5,8 +5,12 @@
  * See LICENSE or go to https://www.gnu.org/licenses/gpl-3.0.txt
  */
 
-import { ReviewStrategy, ScreenName } from '@ulangi/ulangi-common/enums';
-import { ErrorBag, SelectionItem } from '@ulangi/ulangi-common/interfaces';
+import {
+  Feedback,
+  ReviewStrategy,
+  ScreenName,
+} from '@ulangi/ulangi-common/enums';
+import { SelectionItem } from '@ulangi/ulangi-common/interfaces';
 import { ObservableSpacedRepetitionSettingsScreen } from '@ulangi/ulangi-observable';
 import { boundClass } from 'autobind-decorator';
 
@@ -15,23 +19,27 @@ import { SpacedRepetitionSettingsScreenIds } from '../../constants/ids/SpacedRep
 import { LessonScreenStyle } from '../../styles/LessonScreenStyle';
 import { DialogDelegate } from '../dialog/DialogDelegate';
 import { NavigatorDelegate } from '../navigator/NavigatorDelegate';
+import { ReviewFeedbackButtonDelegate } from '../review-feedback/ReviewFeedbackButtonDelegate';
 import { SpacedRepetitionSettingsDelegate } from './SpacedRepetitionSettingsDelegate';
 
 @boundClass
 export class SpacedRepetitionSettingsScreenDelegate {
   private observableScreen: ObservableSpacedRepetitionSettingsScreen;
   private spacedRepetitionSettingsDelegate: SpacedRepetitionSettingsDelegate;
+  private reviewFeedbackButtonDelegate: ReviewFeedbackButtonDelegate;
   private dialogDelegate: DialogDelegate;
   private navigatorDelegate: NavigatorDelegate;
 
   public constructor(
     observableScreen: ObservableSpacedRepetitionSettingsScreen,
     spacedRepetitionSettingsDelegate: SpacedRepetitionSettingsDelegate,
+    reviewFeedbackButtonDelegate: ReviewFeedbackButtonDelegate,
     dialogDelegate: DialogDelegate,
     navigatorDelegate: NavigatorDelegate,
   ) {
     this.observableScreen = observableScreen;
     this.spacedRepetitionSettingsDelegate = spacedRepetitionSettingsDelegate;
+    this.reviewFeedbackButtonDelegate = reviewFeedbackButtonDelegate;
     this.dialogDelegate = dialogDelegate;
     this.navigatorDelegate = navigatorDelegate;
   }
@@ -42,11 +50,14 @@ export class SpacedRepetitionSettingsScreenDelegate {
         initialInterval: this.observableScreen.selectedInitialInterval,
         limit: this.observableScreen.selectedLimit,
         reviewStrategy: this.observableScreen.selectedReviewStrategy,
+        feedbackButtons: this.observableScreen.selectedFeedbackButtons,
       },
       {
-        onSaving: this.showSavingDialog,
-        onSaveSucceeded: this.showSaveSucceededDialog,
-        onSaveFailed: this.showSaveFailedDialog,
+        onSaving: (): void => this.dialogDelegate.showSavingDialog(),
+        onSaveSucceeded: (): void =>
+          this.dialogDelegate.showSaveSucceededDialog(),
+        onSaveFailed: (errorBag): void =>
+          this.dialogDelegate.showSaveFailedDialog(errorBag),
       },
     );
   }
@@ -120,6 +131,42 @@ export class SpacedRepetitionSettingsScreenDelegate {
     );
   }
 
+  public showFeedbackButtonsMenu(
+    valuePairs: readonly [3 | 4 | 5, string][],
+    selectedFeedbackButtons: 3 | 4 | 5,
+    onSelect: (feedbackButtons: 3 | 4 | 5) => void,
+  ): void {
+    this.navigatorDelegate.showSelectionMenu(
+      {
+        items: new Map(
+          valuePairs.map(
+            ([feedbackButtons, feedbackButtonsText]): [
+              number,
+              SelectionItem
+            ] => {
+              return [
+                feedbackButtons,
+                {
+                  testID: SpacedRepetitionSettingsScreenIds.SELECT_FEEDBACK_BUTTONS_BTN_BY_FEEDBACK_BUTTONS(
+                    feedbackButtons,
+                  ),
+                  text: feedbackButtonsText,
+                  onPress: (): void => {
+                    onSelect(feedbackButtons);
+                    this.dialogDelegate.dismiss();
+                  },
+                },
+              ];
+            },
+          ),
+        ),
+        selectedIds: [selectedFeedbackButtons],
+        title: 'Select',
+      },
+      LessonScreenStyle.LIGHT_BOX_SCREEN_STYLES,
+    );
+  }
+
   public showInitialIntervalMenu(
     valuePairs: readonly [number, string][],
     selectedLevel: number,
@@ -156,6 +203,14 @@ export class SpacedRepetitionSettingsScreenDelegate {
     );
   }
 
+  public getButtonsToShow(
+    numberOfFeedbackButtons: 3 | 4 | 5,
+  ): readonly Feedback[] {
+    return this.reviewFeedbackButtonDelegate.getButtonsToShow(
+      numberOfFeedbackButtons,
+    );
+  }
+
   public showIntervalsLightBox(): void {
     this.navigatorDelegate.showLightBox(
       ScreenName.INTERVALS_SCREEN,
@@ -165,26 +220,5 @@ export class SpacedRepetitionSettingsScreenDelegate {
       },
       LessonScreenStyle.LIGHT_BOX_SCREEN_STYLES,
     );
-  }
-
-  private showSavingDialog(): void {
-    this.dialogDelegate.show({
-      message: 'Saving. Please wait...',
-    });
-  }
-
-  private showSaveSucceededDialog(): void {
-    this.dialogDelegate.showSuccessDialog({
-      message: 'Saved successfully.',
-      onClose: (): void => {
-        this.navigatorDelegate.pop();
-      },
-    });
-  }
-
-  private showSaveFailedDialog(errorBag: ErrorBag): void {
-    this.dialogDelegate.showFailedDialog(errorBag, {
-      title: 'SAVE FAILED',
-    });
   }
 }