+++ packages/ulangi-mobile/src/factories/manage/ManageScreenFactory.ts
@@ -33,11 +33,8 @@ import { PrimaryScreenStyle } from '../../styles/PrimaryScreenStyle';
 import { ScreenFactory } from '../ScreenFactory';
 
 export class ManageScreenFactory extends ScreenFactory {
-  public createSetSelectionMenuDelegate(): SetSelectionMenuDelegate {
-    return new SetSelectionMenuDelegate(
-      this.eventBus,
-      this.props.rootStore.setStore,
-      this.createNavigatorDelegate(),
+  public createSetSelectionMenuDelegateWithStyles(): SetSelectionMenuDelegate {
+    return this.createSetSelectionMenuDelegate(
       PrimaryScreenStyle.LIGHT_BOX_SCREEN_STYLES,
     );
   }
@@ -51,7 +48,7 @@ export class ManageScreenFactory extends ScreenFactory {
       PrimaryScreenStyle.LIGHT_BOX_SCREEN_STYLES,
     );
 
-    const setSelectionMenuDelegate = this.createSetSelectionMenuDelegate();
+    const setSelectionMenuDelegate = this.createSetSelectionMenuDelegateWithStyles();
 
     const manageListSelectionMenuDelegate = new ManageListSelectionMenuDelegate(
       navigatorDelegate,