+++ packages/ulangi-mobile/src/factories/spaced-repetition/SpacedRepetitionLessonScreenFactory.ts
@@ -13,6 +13,7 @@ import { AdAfterLessonDelegate } from '../../delegates/ad/AdAfterLessonDelegate'
 import { AdDelegate } from '../../delegates/ad/AdDelegate';
 import { AutoArchiveSettingsDelegate } from '../../delegates/auto-archive/AutoArchiveSettingsDelegate';
 import { ReviewFeedbackBarDelegate } from '../../delegates/review-feedback/ReviewFeedbackBarDelegate';
+import { ReviewFeedbackButtonDelegate } from '../../delegates/review-feedback/ReviewFeedbackButtonDelegate';
 import { ReviewFeedbackDataDelegate } from '../../delegates/review-feedback/ReviewFeedbackDataDelegate';
 import { SpacedRepetitionLessonScreenDelegate } from '../../delegates/spaced-repetition/SpacedRepetitionLessonScreenDelegate';
 import { SpacedRepetitionSaveResultDelegate } from '../../delegates/spaced-repetition/SpacedRepetitionSaveResultDelegate';
@@ -66,10 +67,13 @@ export class SpacedRepetitionLessonScreenFactory extends ScreenFactory {
       autoArchiveSettingsDelegate,
     );
 
+    const reviewFeedbackButtonDelegate = new ReviewFeedbackButtonDelegate();
+
     const reviewFeedbackBarDelegate = new ReviewFeedbackBarDelegate(
       this.observer,
       observableScreen.reviewFeedbackBarState,
       reviewFeedbackDataDelegate,
+      reviewFeedbackButtonDelegate,
     );
 
     const adDelegate = new AdDelegate(