+++ packages/ulangi-mobile/src/factories/theme/ThemeScreenFactory.ts
@@ -37,7 +37,6 @@ export class ThemeScreenFactory extends ScreenFactory {
       observableScreen,
       themeSelectionMenuDelegate,
       dialogDelegate,
-      navigatorDelegate,
     );
   }
 }