+++ packages/ulangi-saga/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.4.0](https://github.com/minhloi/ulangi/compare/v3.3.4...v3.4.0) (2020-02-03)
+
+**Note:** Version bump only for package @ulangi/ulangi-saga
+
+
+
+
+
 ## [3.3.4](https://github.com/minhloi/ulangi/compare/v3.3.3...v3.3.4) (2020-01-25)
 
 **Note:** Version bump only for package @ulangi/ulangi-saga